Add basic script to compare species
This script checks different species to see if they have different names
but the same atom counts. If there is a discrepancy between the two,
this script prompts the user to take a closer look.

This should be improved with chemical graph theory concepts because atom
counts don't accurately determine whether or not two species are the
same.