Remove species ordering in the dataset
When the reactions dataset is created, the reactants and products have
an inherent ordering associated with them. This commit updates the
dataset by duplicating the reactions and swapping the orders of the
reactants and the products. The hope is that the neural network can
learn that there is no preferred order between the reactants and the
products. This will also increase the number of data points to train the
network on.