Allow creation of arbitrary model architectures
ArrheniusNet was created to only allow uniform architectures (specified
number of layers and fixed number of neurons per layer). This commit
extends that functionality to allow the creation of networks that have
an arbitrary number of neurons per layer. It also makes the project
names a tad more descriptive to include the different options being used
to run it.