Attempt to parallelize cross-validation
This commit doesn't produce code that works yet. I tried a bunch of
different approaches using the multiprocessing module and the
concurrent.futures module. However, there is some kind of issue because
some generator expression isn't getting serialized. It seems like it's
much harder to run things in parallel when they are not in the global
scope. There are some workarounds that don't seem as pretty. This
particular attempt seemed like it was almost there but it unfortunately
was not. Maybe I'll get there someday.