Allow completed models to skip cross-validation
Cross-validation takes a lot of time. We don't want to be wasting time
running cross-validation code on models that have already been
processed. This commit makes us smarter in that regard. If a model has
already had cross-validation run on it and we're not forcing a re-run of
cross validation, then this will skip it and move on to the next model
in the list.