Split data into training, testing, and validation
The previous implementation of the code confused the testing and
validation sets (it didn't really have a true testing set). This commit
corrects that mistake (so the data is now correctly referred to as
validation data instead of testing data). This commit also introduces
testing data correctly.

This commit also makes a change to how the model's __call__() is
implemented to allow it to accept Dataset and Subset objects. This
should make it easier to run the model. It also casts the fold indices
to lists so that Subset indexing works as expected.

Finally, this commit also moves where the project names are created to
make things a little more readable and easier to follow, in addition to
a few additional changes.