Skip NA in pandas calculations
By default, pandas aggregation functions (sum, mean, etc) ignore NA's in
calculations instead of returning an NA. While there might be use cases
where this is desirable behavior it was not something that I was
expecting. This commit ensures that these functions return NA if even
one NA value is found in the Series.